Working on test context

+++ src/main/java/com/dataartschool2/stadiumticket/dreamteam/dao/SectorDAO.java
@@ -2,6 +2,6 @@ package com.dataartschool2.stadiumticket.dreamteam.dao;
 
 import com.dataartschool2.stadiumticket.dreamteam.domain.Sector;
 
-public interface SectorDAO extends GenericDAO<Sector, Long>{
+public interface SectorDAO extends GenericDAO<Sector, Integer>{
 	
 }
